Move thread pool size preconditions to build fn (#4)
* Move pool size preconditions to Builder::build.

This fixes the bug that you can't call `fixed_size` with a value
greater than the default for `max_pool_size` (num_cpus). Changing
the order of builder fns in `fixed_size` wouldn't solve the problem,
as then you wouldn't be able to call `fixed_size` with a value
less than the default for `core_pool_size` (also num_cpus).

* Impl Clone and Debug for ThreadPool and Sender

* Add Debug lint.

As part of that, impl Debug for Builder and Config.